Decompose search store
