+++ src/components/Search/store/index.ts
@@ -1,0 +1,3 @@
+export * from "./actions";
+export * from "./reducer";
+export * from "./saga";
